requirements.txt
pandas
pillow
python-docx
flask
requests